# limit the max exec time for nasemi/semi join probe