--- a/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskHelper.h
+++ b/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskHelper.h
@@ -23,7 +23,9 @@
 
 namespace DB
 {
-#define FINISH_STATUS ExecTaskStatus::FINISHED : case ExecTaskStatus::ERROR : case ExecTaskStatus::CANCELLED
+#define FINISH_STATUS                                      \
+    ExecTaskStatus::FINISHED : case ExecTaskStatus::ERROR: \
+    case ExecTaskStatus::CANCELLED
 
 #define UNEXPECTED_STATUS(logger, status) \
     RUNTIME_ASSERT(false, (logger), "Unexpected task status {}", magic_enum::enum_name(status));
@@ -48,7 +50,5 @@ namespace DB
             "Unexpected error reported, detail:\n{}", \
             getCurrentExceptionMessage(true, true));  \
     }
-
-static constexpr int64_t YIELD_MAX_TIME_SPENT_NS = 100'000'000L;
 
 } // namespace DB