--- a/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskTimer.cpp
+++ b/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskTimer.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 PingCAP, Inc.
+// Copyright 2024 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.