--- a/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskTimer.h
+++ b/dbms/src/Flash/Pipeline/Schedule/Tasks/TaskTimer.h
@@ -1,0 +1,36 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <Flash/Pipeline/Schedule/Tasks/TaskProfileInfo.h>
+#include <common/types.h>
+
+namespace DB
+{
+static constexpr int64_t YIELD_MAX_TIME_SPENT_NS = 100'000'000L;
+
+struct TaskTimer
+{
+    TaskProfileInfo & profile_info;
+    UInt64 current_exec_time = 0;
+    UInt64 updateCurrentExecTime()
+    {
+        current_exec_time += profile_info.elapsedFromPrev();
+        return current_exec_time;
+    }
+};
+
+extern thread_local TaskTimer * current_task_timer;
+} // namespace DB