--- a/dbms/src/Flash/tests/gtest_join_executor.cpp
+++ b/dbms/src/Flash/tests/gtest_join_executor.cpp
@@ -18,6 +18,10 @@
 
 namespace DB
 {
+namespace FailPoints
+{
+extern const char force_semi_join_time_exceed[];
+} // namespace FailPoints
 namespace tests
 {
 class JoinExecutorTestRunner : public DB::tests::JoinTestRunner
@@ -2864,6 +2868,7 @@ try
 {
     using tipb::JoinType;
     std::vector<UInt64> cross_join_shallow_copy_thresholds{1, DEFAULT_BLOCK_SIZE * 100};
+    std::vector<bool> semi_join_time_exceed = {true, false};
     /// One join key(t.a = s.a) + no other condition.
     /// left table(t) + right table(s) + result column.
     const std::vector<std::tuple<ColumnsWithTypeAndName, ColumnsWithTypeAndName, ColumnWithTypeAndName>> t1
@@ -2915,7 +2920,18 @@ try
                 = context.scan("null_aware_semi", "t")
                       .join(context.scan("null_aware_semi", "s"), type, {col("a")}, {}, {}, {}, {}, 0, is_null_aware)
                       .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune
                 = context.scan("null_aware_semi", "t")
                       .join(context.scan("null_aware_semi", "s"), type, {col("a")}, {}, {}, {}, {}, 0, is_null_aware)
@@ -3018,7 +3034,18 @@ try
                                    0,
                                    is_null_aware)
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune = context.scan("null_aware_semi", "t")
                                             .join(
                                                 context.scan("null_aware_semi", "s"),
@@ -3130,7 +3157,18 @@ try
                                    0,
                                    is_null_aware)
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune = context.scan("null_aware_semi", "t")
                                             .join(
                                                 context.scan("null_aware_semi", "s"),
@@ -3298,7 +3336,18 @@ try
                                    0,
                                    is_null_aware)
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune = context.scan("null_aware_semi", "t")
                                             .join(
                                                 context.scan("null_aware_semi", "s"),
@@ -3408,7 +3457,18 @@ try
                                    0,
                                    is_null_aware)
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune = context.scan("null_aware_semi", "t")
                                             .join(
                                                 context.scan("null_aware_semi", "s"),
@@ -3509,7 +3569,18 @@ try
                                    0,
                                    is_null_aware)
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
             auto request_column_prune = context.scan("null_aware_semi", "t")
                                             .join(
                                                 context.scan("null_aware_semi", "s"),
@@ -3565,6 +3636,7 @@ CATCH
 TEST_F(JoinExecutorTestRunner, SemiJoin)
 try
 {
+    std::vector<bool> semi_join_time_exceed = {true, false};
     using tipb::JoinType;
     /// One join key(t.a = s.a) + no other condition.
     /// left table(t) + right table(s) + result column.
@@ -3608,7 +3680,18 @@ try
         {
             auto reference = genSemiJoinResult(type, left, res);
             auto request = context.scan("semi", "t").join(context.scan("semi", "s"), type, {col("a")}).build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
         }
     }
 
@@ -3652,7 +3735,18 @@ try
                 = context.scan("semi", "t")
                       .join(context.scan("semi", "s"), type, {col("a")}, {}, {}, {lt(col("t.c"), col("s.c"))}, {})
                       .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
         }
     }
 
@@ -3696,7 +3790,18 @@ try
             auto request = context.scan("semi", "t")
                                .join(context.scan("semi", "s"), type, {col("a"), col("b")}, {})
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
         }
     }
 
@@ -3778,7 +3883,18 @@ try
                                    {lt(col("t.c"), col("s.c"))},
                                    {})
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
         }
     }
 
@@ -3810,7 +3926,18 @@ try
             auto request = context.scan("semi", "t")
                                .join(context.scan("semi", "s"), type, {col("a"), col("b")}, {})
                                .build(context);
-            executeAndAssertColumnsEqual(request, reference);
+            for (auto need_force_semi_join_time_exceed : semi_join_time_exceed)
+            {
+                if (need_force_semi_join_time_exceed)
+                {
+                    FailPointHelper::enableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                else
+                {
+                    FailPointHelper::disableFailPoint(FailPoints::force_semi_join_time_exceed);
+                }
+                executeAndAssertColumnsEqual(request, reference);
+            }
         }
     }
 }